add interrupt tdd test

+++ src/SPITemp.c
@@ -3,9 +3,9 @@
 #include "Exception.h"
 #include "CException.h"
 
-int calculateTemp(uint8_t	*TempData)
+float calculateTemp(uint8_t	*TempData)
 {
-	int temp=0;
+	float temp=0;
 
 	if (TempData[0] & 0x4) {
 		throwException(THERMOCOUPLE_INPUT_OPEN,NULL,"bit 2 = 1, thermocouple input is open.");
@@ -13,7 +13,7 @@ int calculateTemp(uint8_t	*TempData)
 	else{
 		//each point equal to 0.25 celsius
     //data is store from bit 14 - bit 3
-		temp = (int)(((TempData[1]<<5)|(TempData[0]>>3))*0.25);
+		temp = (float)(((TempData[1]<<5)|(TempData[0]>>3))*0.25);
 	}
   return temp;
 }
